refactor: extract shared tool internals to tools/core.py (Task 18)
Move EventEmitterBridge, response helpers, UPS client cache, row
normalization, and preview enrichment from tools_v2.py into the new
tools/core.py submodule. tools_v2 re-imports all names for backward
compatibility. Update test patch targets to tools.core namespace.
