+++ src/orchestrator/agent/tools/__init__.py
@@ -1,0 +1,6 @@
+"""Agent tool subpackage — shared internals in core.py.
+
+Tool handler submodules import shared helpers from tools.core.
+The get_all_tool_definitions() registry remains in tools_v2.py
+until handler migration (Task 19) is complete.
+"""
